core/cmms/requirements.txt
# ChatterFix CMMS - Optimized Cloud Run Requirements
# Essential dependencies for stable Cloud Run deployment

# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# HTTP client for AI API calls
httpx>=0.25.0
requests>=2.31.0

# Authentication and security
pyjwt>=2.8.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Template and static files
jinja2>=3.1.2
aiofiles>=23.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Basic data processing
pandas>=2.1.0
openpyxl>=3.1.0
numpy>=1.25.0

# Monitoring and logging
prometheus-client>=0.17.0

# Basic AI features (removed heavy ML dependencies for stability)
pillow>=10.0.0

# Real-time WebSocket support
websockets>=11.0.0

# System monitoring
psutil>=5.9.0

# PostgreSQL database support
psycopg2-binary>=2.9.0